Added a functionality:
queue_manager is now aware of how many times the
host_queue has been cycled.

The aim for this was to force the queue_manager to cycle through the current queue at least once and finish the current cycle before filling the host_queue with a new list.

The idea works as follows:

put_new_list() is called
- check if queue has been cycled at least once and new cycle has not started
- if true then call empty_queue, add new_list to the host_queue, reset "_times_fully_parsed" variable to 0 and
   "_new_list_holder" to none,
   else keep new_list in new_list_holder variable and return finishing the current cycle

next_host() is called
- check if _new_list_holder is none
   if true continue, else call put_new_list()
- get first element
- increment "_counter" variable
- append current element at the end of host_queue
- check if counter equals the length of the host_queue
 - if true reset "_counter" to 0 and increment "_times_fully_parsed" variable
    return current element