First part of a bug fix to a multi layered Problem.
This fix is focusing on a race condition during start up.

Problem description:
All modules in crawler are started by hotcat.py in separated processes.
Currently the queue manager checks and loads a dump file prior being registered to the queue server process.
Meanwhile the other modules try to connect to the queue manager trough the queue server and fail due the queue manager
is still occupied with the dump file.

Fix:
Load the dump file in a separate thread.